+++ peer.go
@@ -1,0 +1,219 @@
+package main
+
+import (
+	"bytes"
+	"context"
+	"crypto/rand"
+	"encoding/json"
+	"github.com/libp2p/go-libp2p"
+	"github.com/libp2p/go-libp2p-core/crypto"
+	"github.com/multiformats/go-multiaddr"
+	"github.com/pkg/errors"
+	"github.com/sirupsen/logrus"
+	"io/ioutil"
+	"net/http"
+	"strings"
+)
+
+type trackerAnnounceRequest struct {
+	IP        string `json:"ip"`
+	Port      string `json:"port"`
+	Multiaddr string `json:"multiaddr"`
+	Action    string `json:"action"`
+}
+
+type trackerAnnounceResponseData struct {
+	ShareID string `json:"shareId"`
+}
+
+type trackerAnnounceResponse struct {
+	Action string                      `json:"action"`
+	Data   trackerAnnounceResponseData `json:"data"`
+}
+
+type trackerPulseRequest struct {
+	Action string            `json:"action"`
+	Data   map[string]string `json:"data"`
+}
+
+type trackerPulseResponse struct {
+	Action string `json:"action"`
+}
+
+type trackerPushRequest struct {
+	ShareID string `json:"shareId"`
+	Action  string `json:"action"`
+}
+
+type trackerPushResponseData struct {
+	Multiaddr string `json:"multiaddr"`
+}
+
+type trackerPushResponse struct {
+	Action string                  `json:"action"`
+	Data   trackerPushResponseData `json:"data"`
+}
+
+type sobaniPeer struct {
+	IP        string
+	Port      string
+	Multiaddr string
+}
+
+func newSobaniPeer() (*sobaniPeer, error) {
+	peer := &sobaniPeer{}
+
+	// Creates a new RSA key pair for this host.
+	var r = rand.Reader
+	prvKey, _, err := crypto.GenerateKeyPairWithReader(crypto.RSA, 2048, r)
+	if err != nil {
+		panic(err)
+	}
+
+	// 0.0.0.0 will listen on any interface device.
+	sourceMultiAddr, _ := multiaddr.NewMultiaddr("/ip4/0.0.0.0/tcp/0")
+
+	// libp2p.New constructs a new libp2p Host.
+	// Other options can be added here.
+	host, err := libp2p.New(
+		context.Background(),
+		libp2p.ListenAddrs(sourceMultiAddr),
+		libp2p.Identity(prvKey),
+	)
+	if err != nil {
+		logrus.Fatal(err)
+		return nil, err
+	}
+
+	// Set a function as stream handler.
+	// This function is called when a peer connects, and starts a stream with this protocol.
+	// Only applies on the receiving side.
+	host.SetStreamHandler("/sobani/1.0.0", handleStream)
+
+	// Let's get the actual TCP port from our listen multiaddr, in case we're using 0 (default; random available port).
+	var port string
+	for _, la := range host.Network().ListenAddresses() {
+		if p, err := la.ValueForProtocol(multiaddr.P_TCP); err == nil {
+			port = p
+			break
+		}
+	}
+
+	if port == "" {
+		return nil, errors.New("was not able to find actual local port")
+	}
+	peer.Port = port
+	peer.Multiaddr = host.ID().Pretty()
+
+	publicIP, err := getPublicIP()
+	if err != nil {
+		logrus.Debug(err)
+		return nil, err
+	}
+	peer.IP = *publicIP
+
+	return peer, nil
+}
+
+// This function announce self to a public tracker
+// It returns the corresponding share ID generated by tracker and any error encountered.
+func (peer *sobaniPeer) announceToTracker(trackerURL *string) (*trackerAnnounceResponse, error) {
+	logrus.Info("Announcing to tracker at: %s", *trackerURL)
+
+	request := &trackerAnnounceRequest{
+		IP:        peer.IP,
+		Port:      peer.Port,
+		Multiaddr: peer.Multiaddr,
+		Action:    "announce",
+	}
+
+	requestJSON, _ := json.Marshal(request)
+	logrus.Debug("JSON: ", string(requestJSON))
+
+	body, err := postToTracker(&requestJSON, trackerURL)
+	if err != nil {
+		return nil, err
+	}
+	logrus.Debug("response Body:", string(*body))
+
+	res := trackerAnnounceResponse{}
+	err = json.Unmarshal(*body, &res)
+	if err != nil {
+		return nil, err
+	}
+
+	return &res, nil
+}
+
+// This function get peer info by share ID on public tracker
+// It returns the requested peer info on tracker and any error encountered.
+func (peer *sobaniPeer) getPeerInfo(shareID *string, trackerURL *string) (*trackerPushResponse, error) {
+	logrus.Infof("Connecting to %s via %s", *shareID, *trackerURL)
+
+	request := &trackerPushRequest{
+		ShareID: *shareID,
+		Action:  "push",
+	}
+
+	requestJSON, _ := json.Marshal(request)
+	logrus.Debug("JSON: ", string(requestJSON))
+
+	body, err := postToTracker(&requestJSON, trackerURL)
+	if err != nil {
+		return nil, err
+	}
+	logrus.Debug("response Body:", string(*body))
+
+	res := trackerPushResponse{}
+	err = json.Unmarshal(*body, &res)
+	if err != nil {
+		return nil, err
+	}
+
+	return &res, nil
+}
+
+// This function will try to get system's public IP address via ip.cip.cc
+// It returns the one of public IP addresses and any error encountered.
+func getPublicIP() (*string, error) {
+	logrus.Info("Finding public IP address...")
+	res, err := http.Get("http://ip.cip.cc/")
+	if err != nil {
+		logrus.Fatal(err)
+		return nil, err
+	}
+	defer res.Body.Close()
+
+	if res.Status != "200 OK" {
+		logrus.Fatal(err)
+		return nil, err
+	}
+
+	body, _ := ioutil.ReadAll(res.Body)
+	ip := strings.TrimSpace(string(body))
+	return &ip, nil
+}
+
+func postToTracker(data *[]byte, trackerURL *string) (*[]byte, error) {
+	req, err := http.NewRequest("POST", *trackerURL, bytes.NewBuffer(*data))
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Set("Content-Type", "application/json;charset=utf-8")
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return nil, err
+	}
+
+	if resp.Status != "200 OK" {
+		logrus.Error("Tracker error")
+	}
+
+	body, err := ioutil.ReadAll(resp.Body)
+	if err != nil {
+		return nil, err
+	}
+	return &body, nil
+}
