+++ sobani.go
@@ -2,63 +2,12 @@ package main
 
 import (
 	"bufio"
-	"bytes"
-	"context"
-	"crypto/rand"
-	"encoding/json"
 	"flag"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"net/http"
-	"os"
-	"strings"
-
-	"github.com/libp2p/go-libp2p"
-	"github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/libp2p/go-libp2p-core/network"
-	"github.com/multiformats/go-multiaddr"
 	"github.com/sirupsen/logrus"
+	"os"
 )
-
-type trackerAnnounceRequest struct {
-	IP        string `json:"ip"`
-	Port      string `json:"port"`
-	Multiaddr string `json:"multiaddr"`
-	Action    string `json:"action"`
-}
-
-type trackerAnnounceResponseData struct {
-	ShareID string `json:"shareId"`
-}
-
-type trackerAnnounceResponse struct {
-	Action string                      `json:"action"`
-	Data   trackerAnnounceResponseData `json:"data"`
-}
-
-type trackerPulseRequest struct {
-	Action string            `json:"action"`
-	Data   map[string]string `json:"data"`
-}
-
-type trackerPulseResponse struct {
-	Action string `json:"action"`
-}
-
-type trackerPushRequest struct {
-	ShareID string `json:"shareId"`
-	Action  string `json:"action"`
-}
-
-type trackerPushResponseData struct {
-	Multiaddr string `json:"multiaddr"`
-}
-
-type trackerPushResponse struct {
-	Action string                  `json:"action"`
-	Data   trackerPushResponseData `json:"data"`
-}
 
 func handleStream(s network.Stream) {
 	logrus.Println("Got a new stream!")
@@ -103,72 +52,20 @@ func writeData(rw *bufio.ReadWriter) {
 	}
 }
 
-// This function will try to get system's public IP address via ip.cip.cc
-// It returns the one of public IP addresses and any error encountered.
-func getPublicIP() (*string, error) {
-	logrus.Info("Finding public IP address...")
-	res, err := http.Get("http://ip.cip.cc/")
-	if err != nil {
-		logrus.Fatal(err)
-		return nil, err
+// This function setup the log style and log level
+func setupLog(debug *bool) {
+	// setup log
+	logrus.SetOutput(os.Stdout)
+	Formatter := &logrus.TextFormatter{
+		EnvironmentOverrideColors: true,
+		FullTimestamp:             true,
+		TimestampFormat:           "2020-02-27 00:43:00",
 	}
-	defer res.Body.Close()
-
-	if res.Status != "200 OK" {
-		logrus.Fatal(err)
-		return nil, err
+	logrus.SetFormatter(Formatter)
+	logrus.SetLevel(logrus.InfoLevel)
+	if *debug {
+		logrus.SetLevel(logrus.DebugLevel)
 	}
-
-	body, _ := ioutil.ReadAll(res.Body)
-	ip := strings.TrimSpace(string(body))
-	return &ip, nil
-}
-
-// This function announce self to a public tracker
-// It returns the corresponding share ID generated by tracker and any error encountered.
-func announceToTracker(trackerURL *string, request *trackerAnnounceRequest) (*trackerAnnounceResponse, error) {
-	logrus.Info("Announcing to tracker at: %s", *trackerURL)
-
-	requestJSON, _ := json.Marshal(request)
-	logrus.Debug("JSON: ", string(requestJSON))
-	req, err := http.NewRequest("POST", *trackerURL, bytes.NewBuffer(requestJSON))
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Content-Type", "application/json;charset=utf-8")
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		return nil, err
-	}
-
-	if resp.Status != "200 OK" {
-		logrus.Error("Tracker error")
-	}
-
-	body, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, err
-	}
-	logrus.Debug("response Body:", string(body))
-
-	res := trackerAnnounceResponse{}
-	err = json.Unmarshal(body, &res)
-	if err != nil {
-		return nil, err
-	}
-
-	return &res, nil
-}
-
-// This function get peer info by share ID on public tracker
-// It returns the requested peer info on tracker and any error encountered.
-func getPeerInfo(request *trackerPushRequest, trackerURL *string) (*trackerPushResponse, error) {
-	logrus.Infof("Connecting to %s via %s", (*request).ShareID, *trackerURL)
-	// todo: post "push" action to tracker
-
-	return nil, nil
 }
 
 func main() {
@@ -184,90 +81,25 @@ func main() {
 		os.Exit(0)
 	}
 
-	// setup log
-	logrus.SetOutput(os.Stdout)
-	Formatter := &logrus.TextFormatter{
-		EnvironmentOverrideColors: true,
-		FullTimestamp:             true,
-		TimestampFormat:           "2020-02-27 00:43:00",
-	}
-	logrus.SetFormatter(Formatter)
-	logrus.SetLevel(logrus.InfoLevel)
-	if *debug {
-		logrus.SetLevel(logrus.DebugLevel)
-	}
-
-	// Creates a new RSA key pair for this host.
-	var r io.Reader = rand.Reader
-	prvKey, _, err := crypto.GenerateKeyPairWithReader(crypto.RSA, 2048, r)
+	setupLog(debug)
+	peer, err := newSobaniPeer()
 	if err != nil {
+		// todo: retry
 		panic(err)
 	}
-
-	// 0.0.0.0 will listen on any interface device.
-	sourceMultiAddr, _ := multiaddr.NewMultiaddr("/ip4/0.0.0.0/tcp/0")
-
-	// libp2p.New constructs a new libp2p Host.
-	// Other options can be added here.
-	host, err := libp2p.New(
-		context.Background(),
-		libp2p.ListenAddrs(sourceMultiAddr),
-		libp2p.Identity(prvKey),
-	)
+	res, err := peer.announceToTracker(trackerURL)
 	if err != nil {
+		// todo: retry
 		panic(err)
 	}
 
 	if *connect == "" {
-		// Set a function as stream handler.
-		// This function is called when a peer connects, and starts a stream with this protocol.
-		// Only applies on the receiving side.
-		host.SetStreamHandler("/chat/1.0.0", handleStream)
-
-		// Let's get the actual TCP port from our listen multiaddr, in case we're using 0 (default; random available port).
-		var port string
-		for _, la := range host.Network().ListenAddresses() {
-			if p, err := la.ValueForProtocol(multiaddr.P_TCP); err == nil {
-				port = p
-				break
-			}
-		}
-
-		if port == "" {
-			panic("was not able to find actual local port")
-		}
-
-		publicIP, err := getPublicIP()
-		if err != nil {
-			logrus.Debug(err)
-		}
-		request := &trackerAnnounceRequest{
-			IP:        *publicIP,
-			Port:      port,
-			Multiaddr: host.ID().Pretty(),
-			Action:    "announce",
-		}
-		res, err := announceToTracker(trackerURL, request)
-		if err != nil {
-			// todo: retry
-			panic(err)
-		}
-
 		logrus.Info("Run './chat -connect %s -tracker %s' on another console.\n", res.Data.ShareID, trackerURL)
 
 		// Hang forever
 		<-make(chan struct{})
 	} else {
-		logrus.Debug("This node's multiaddresses:")
-		for _, la := range host.Addrs() {
-			logrus.Debug(" - %v\n", la)
-		}
-
-		request := &trackerPushRequest{
-			ShareID: *connect,
-			Action:  "push",
-		}
-		info, err := getPeerInfo(request, trackerURL)
+		info, err := peer.getPeerInfo(connect, trackerURL)
 		if err != nil {
 			logrus.Errorf("Cannot get peer `%s` from %s", *connect, *trackerURL)
 		} else {
